Refactor craftBidAdapter (#12517)
